refactor loader name
