+++ flutter/lib/src/ffplay_session.dart
@@ -24,7 +24,7 @@ import 'package:ffi/ffi.dart';
 
 import '../ffmpeg_kit_extended_flutter.dart';
 import 'callback_manager.dart';
-import 'ffmpeg_kit_flutter_loader.dart';
+import 'ffmpeg_kit_extended_flutter_loader.dart';
 
 /// Session for playing media using FFplay.
 ///
